pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "derbyfish-flow"
version = "1.0.0"
description = "DerbyFish Flow API - blockchain infrastructure for verified Fish Cards"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Flask>=2.3.3",
    "supabase>=2.8.1",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "flow-py-sdk>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
]

[project.scripts]
derbyfish-flow-cli = "cli:main"

[tool.setuptools.packages.find]
where = ["src/python"]
include = ["cli*"]
